Alterado variavel price de Double para BigDecimal

+++ src/test/java/com/eliseu/BikesRegistration/BikeControllerAuthInMemoryTest.java
@@ -23,6 +23,7 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -53,7 +54,7 @@ public class BikeControllerAuthInMemoryTest {
 
     @Before
     public void setup() {
-        Bike bike = new Bike(1L, "Before's Bike", "BB", 2598.3, LocalDate.of(2014, 10, 26), "Jonh", "Bikes Store");
+        Bike bike = new Bike(1L, "Before's Bike", "BB", BigDecimal.valueOf(2598.30), LocalDate.of(2014, 10, 26), "Jonh", "Bikes Store");
         BDDMockito.when(bikeRepository.findById(bike.getId())).thenReturn(java.util.Optional.of(bike));
     }
 
@@ -75,8 +76,8 @@ public class BikeControllerAuthInMemoryTest {
 
     @Test
     public void listBikesWhenUsernameAndPasswordAreCorrectShouldReturnStatusCode200() {
-        List<Bike> bikes = asList(new Bike(2L, "Jonh's Bike", "Caloi", 850.30, LocalDate.of(2012, 12, 16), "Jonh", "Bikes Store"),
-                new Bike(3L, "Marias's Bike", "MonarkCeci", 1350.55, LocalDate.of(2014, 10, 26), "Maria", "Bikes Store"));
+        List<Bike> bikes = asList(new Bike(2L, "Jonh's Bike", "Caloi", BigDecimal.valueOf(850.30), LocalDate.of(2012, 12, 16), "Jonh", "Bikes Store"),
+                new Bike(3L, "Marias's Bike", "MonarkCeci", BigDecimal.valueOf(1350.55), LocalDate.of(2014, 10, 26), "Maria", "Bikes Store"));
         BDDMockito.when(bikeRepository.findAll()).thenReturn(bikes);
         ResponseEntity<String> response = restTemplate.getForEntity("/v1/bikes", String.class);
         Assertions.assertThat(response.getStatusCodeValue()).isEqualTo(200);
@@ -125,7 +126,7 @@ public class BikeControllerAuthInMemoryTest {
 
     @Test
     public void createWhenDescriptionIsNullShouldReturnStatusCode400BadRequest() {
-        Bike bike = new Bike(3L, "", "MonarkCeci", 1350.55, LocalDate.of(2014, 10, 26), "Maria", "Bikes Store");
+        Bike bike = new Bike(3L, "", "MonarkCeci", BigDecimal.valueOf(1350.55), LocalDate.of(2014, 10, 26), "Maria", "Bikes Store");
         BDDMockito.when(bikeRepository.save(bike)).thenReturn(bike);
         ResponseEntity<String> response = restTemplate.postForEntity("/v1/bikes/", bike, String.class);
         Assertions.assertThat(response.getStatusCodeValue()).isEqualTo(400);
@@ -134,7 +135,7 @@ public class BikeControllerAuthInMemoryTest {
 
     @Test
     public void createShouldPersistDataAndReturnStatusCode201() throws Exception {
-        Bike bike = new Bike("Cecizinha", "MonarkCeci", 1350.55, LocalDate.of(2014, 10, 26), "Maria", "Bikes Store");
+        Bike bike = new Bike("Cecizinha", "MonarkCeci", BigDecimal.valueOf(1350.55), LocalDate.of(2014, 10, 26), "Maria", "Bikes Store");
         BDDMockito.when(bikeRepository.save(bike)).thenReturn(bike);
         ResponseEntity<Bike> responseTemp = restTemplate.getForEntity("/v1/bikes/3", Bike.class);//testestestesa
         System.out.println(responseTemp);
